// src/employee/enums/employee.enum.ts
export enum EmployeeMessages {
    NOT_FOUND = 'Employee not found',
    ALREADY_EXISTS = 'Employee already exists',
    CREATED = 'Employee session created successfully',
    UPDATED = 'Employee sesion updated successfully',
    DELETED = 'Employee deleted successfully',
    EMAIL_ALREADY_REGISTERED = 'Email already registered',
}